Stricter checking for ambiguous rules
Check that all items in alternatives have non-overlapping first sets. Two
sets are overlapping if they both contain an entry pointing to the same
struct literal or if the contain entries where the literals have the same
patterns (as strings)

This still misses the case where literals have equivalent regular
expresions (e.g. /[ab]/ in one and /[ba]/ in the other) but that seems very
hard to determine.