Provider that uses specs
The spec provider is hooked into libaugeas and loaded. It reads spec files
from /usr/share/augeas/spec and paths mentioned on AUGEAS_SPECLIB. Files
are parsed according to the specs loaded, and values are written into the
tree after parsing those files.

Still missing is saving of the tree back to file, and in particular adding
new entries in the tree to such files.