Build an explicit AST
Instead of representing the parsed file by a simple list of tokens, build a
fullblown AST for that file. The AST is actually not very abstract, and is
directly the parse tree from parsing a file with a certain grammar.

Split parsing and evaluating into to passes; references to fields
are too hard to resolve with a partially constructed AST.

Also adds a routine (ast_dot) to output the AST as a dot file; this eases
debugging tremendously. augparse will now produce dot files when given '-P
ast'

Remove aug_token and any mention of it - that has been replaced by struct
ast.