Emit a the current tree into a file
Based on the AST, generate the file that lead to that AST. Changes to the
config tree are mirrored by corresponding changes in the AST before it is
written to disk.

This is a first version that works for some restricted set of grammars, but
needs a lot more testing.