Only save files with actual changes
Split emitting into two parts: updating the AST from the config tree and
actually writing the AST to file. do_insertion and do_deletion now report
if they made any changes to the tree so that the provider can decide
whether to save the file or not.

As a byproduct, parts of the AST that are no longer needed are actually
freed.