For quantified groups, count the quantifier as the group
Previously, for a match like '(A)*' A was counted as a group, which meant
that if an action was attached to the group, all the children of the '*'
node in the AST had the same path. By counting the '*' as the group, not
'A', the children will have different name if different subtrees are
produced in the group.

This ensures that in the AST, the children of '*' and '+' have different
names and simplifies the 'put' direction of the transformation
considerably.