Distribute enters so we have one per match
Simultaneous traversal of the AST and the config tree is much simpler if
any match will only go at most one level deeper. That guarantees that
entering a child in the config tree corresponds to entering an AST node
(possibly followed by entering some more AST nodes labelled with the same
path)

To this end, split actions with multiple enters into fake SEQUENCE matches
(followed by the match that held the initial action). The fake SEQUENCE
matches all have exactly one child in their matches field, and one of the
enter instructions from the action for them.

Actions are also lifted up in the grammar as far as possible; in
particular, this pulls constants (but not $seq) out of iterators. This
ensures that the children of an AST node differ in exactly one path
component.