Respect ordering of nodes in the config tree.
Previously, reordering of branches in the config tree did not lead to a
corresponding reordering of elements in the file. Branches in the AST are
now reordered in such a way that their ordering reflects the ordering in
the config tree.

Iterator nodes with a $seq action need to be treated special: they do not
change the path for the current (iterator) node, only for the child nodes.
That is terribly ugly.