Initialize fields used as path components
When a field is used in a path component, initialize its value from the
path. Since it is not possible to change an existing path, this operation
is only needed when new parts of the AST are created.

The code assumes that the leaf for the path component is in the subtree
where the leaf's value is used. That is not necessarily true; at least,
it's not being checked from the grammar.