For split nodes, append entries at the last occurence
It is possible that one logical node is split across several physical
nodes to accomodate settings that are spread across several lines like
'AcceptEnv' for sshd_config. Such lines lead to multiple nodes
/system/config/sshd/AcceptEnv (all siblings of each other)

When a call to aug_set creates a new entry underneath such a split node,
make sure that that the new node is put under the last such node. This
helps maintain the illusion for users that the split node is really just
one node.