Fix nested splits
For a lens like a . (b . c)*, tree_split used to return a split that looked
like [a; b; c; b; c] That is wrong: when we split on the outer '.', we need
to produce a split with exactly two entries: [a; b] (where b stands as the
first entry in the tree that '*' will act on.

split_tree_car cleans up recursively computed splits in this manner; those
subsplits are mostly computed for their sideeffect, particularly for the
fact that they change state->tree.