Base parsing completely off regexp matching.
Parsing now uses regexp matching, and not the CFG-style parsing based on
first sets previously used. This actually makes it easier to write specs,
since parsing decisions are made based on the complete string, not just the
first 'token'.

The downside is that rules can not be recursive anymore, as that would
make the matched language non-regular.

The main motivator for this change is that it makes it possible to do
stricter checking of the input spec using type checking similar to
boomerang's string lenses based on operations on regular languages.

Unfortunately, we have to keep first/follow computations to construct the
fairly tricky regexps that are used for ANY patterns.