Allow influencing the behavior of save
By passing one of the flags AUG_SAVE_BACKUP or AUG_SAVE_NEWFILE to
aug_init, it is possible to influence the behavior of aug_save to either
overwrite files in palce, save originals with extension .augsave, or put
changes into files with extension .augnew.

Also added switches to augtool to set this from the command line, and to
set the filesystem root.