Add typechecking of lenses
Concatenation and iteration of lenses is now checked for ambiguity. Union
is checked for disjointness, but that is only flagged as a warning, since
it's mostly a harmless problem, and fixing it requires writing nasty
regexps in the spec files.

Fix existing specs to pass typeckecking. Typechecking is extremely slow
right now, some of that will be addressed shortly. For now, augtool does
not do typechecking, but augparse does by default.