When splitting the tree, check that we are atthe right subtree
Without the check, spliting at a subtree just gobbles up trees without any
regard for what that tree is. This becmoes a problem in a sequence like
  [ COMMENT ] * . [ key FOO . store BAR ] *
when there is no comment, and therefore no subtree witha NULL label. A tree
with a single node matching FOO is wrongly split as if there was a COMMENT
(which, if it were there would have to have a NULL label)