Allow rm from toplevel of tree
The only issue with this is that the dirty flag can not be set properly
when deleting from the toplevel of the tree since the flag has to be set on
the parent, which we don't have.

This is really only an issue if (a) we delete the whole tree (b) the next
sibling of the node we delete is semantically different from the deleted
node (e.g. because it goes into a different file)

Either way, this only happens during tests; in the 'real' tree, toplevel
nodes are fixed, like /files and /augeas, and we don't really have an issue
there.