Clean up some confusion around the semantics of store
Store does not create a tree; instead it puts a value in place where the
enclosing subtree will find it.

During lens building, there is now a check to make sure the key and value
are set in at most one place in each subtree. Ideally, this check would
happen in the typechecker, but it would require that we parametrize the
lens type, considerably complicating matters.