Fix lens selection in union
The logic in put to select the appropriate branch in unions got it wrong if
the first branch matched the empty word, because applies only checked that
the atype for the sublens matched some of the tree. It has to check for a
complete match of its tree.

This change fixes that and adds a test to demonstrate the problem.