Remove bad assertions
It is perfectly legal for create_* functions to be called with a non-NULL
skeleton; that happens for example when the union combinator puts a tree
that "jumped branches", i.e. whose get used one branch of the union, and
the put uses the other branch because of changes to the tree.