Report errors more clearly in the tree
When an error happens during get/put, store additional info in the tree:
for get errors, store the position in the file where the error occured, for
put errors, the path of the node. In both cases, also store the
human-readable message.

For a file F, all this information is stored underneath /augeas/F/error,
with the value associated with /augeas/F/error giving some indication about
the kind of error encountered.