Fix a major bug in the get implementation of lenses
We used to process the input from left to right, matching the regular
expressions of lenses with the input successively. That is very wrong, in
particular, it leads to lens union choosing the wrong branch: in the union
l1|l2, union would always choose l1 if l1 matched a prefix of l2.

We now process the input in a top-down fashion, splitting the string into
parts as we descend in the tree of lenses. As evidenced by the test
pass_prefix_union, that fixes the bad behavior.

Incidentally, it also uncovered another bug in the test_hosts.aug, where a
test passed, even though the string we were using did not end with the
newline that is mandatory for Hosts.record.

This change also does away with pasing flags and a FILE* for logging
details about hte parse process; they don't really make sense anymore, nor
is that behavior very useful in a library.