Produce more helpful errors
The previous check whether we could proces the entire input was too crude
and produced the unhelpful error message "Get could not process entire
input", pointing to the start of the string, even if the problem was much
farther down the line.

With this change, error messages contain an indication to where the error
happened that is a little closer to the actual trouble spot.