Canonicalize file names before writing
We used to clobber symlinks because writing a file consists of writing to
an intermediate .augnew file, and then moving that over the target file. To
keep target files that are symlinks intact, we now canonicalize the target
filename and then move the intermediate file to the canonicalized target
file.