Check that 'put' fails when invalid tree entries are present
put_quant_star and create_quant_star did not compare the part of the tree
they should transform against the part of the tree they actually
transformed. This lead to parts of the tree getting ignored during 'put' if
the tree contained invalid nodes, and ultimately to files getting
truncated.

This is now checked and causes a 'Short iteration' error, avoiding file
truncation. The fix also uncovered an error in the test for Aptsources.lns,
where an invalid tree node was put into the tree, but did not lead to a
test failure.

Fixes bug https://fedorahosted.org/augeas/ticket/1