Avoid O(n^2) runtime by appending to lists in constant time
The data structure used to keep track of successive matches of an iterated
lens, a split list, was being appended to by traversing the list for each
new element. Now, we remember the last element on the list and append in
constant time.

The same was true in a few other places: the building of a tree and of
skeletons suffered from the same problem.

Dicts still have that problem - they need to be changed to hash tables to
address this issue.