Posix ERE compliant escaping
Posix extended regular expressions treat a backslash outside of a character
class always as an escape character, with the character sequence '\\x'
being identical to 'x', even if 'x' is a special char like '['.

Inside character classes, the backslash does not escape anything, and is
treated as an ordianry character.

See regex(7) or
http://www.opengroup.org/onlinepubs/009695399/basedefs/xbd_chap09.html#tag_09_04