aug_mv: fix bug when moving at the root level
When moving from the root level, the source node needs to be removed from
aug->tree, not aug->tree->children. In addition, the list_remove macro must
be passed aug->tree and not a local variable holding that value, since we
might have to remove the list entry that aug->tree is pointing to.

Added test in test-mv.sh to test that moving at the root level works.