Typecheck del lens
The builtin lens 'del RE STR' was missing a typecheck: the default string
STR must match the regular expression RE, otherwise we might produce text
that we cant parse back in.

This patch fixes that, adds a test that the check is performed, and fixes
all the places in existing lenses where a use of 'del' failed the
typecheck.