Properly typecheck the '?' operator
The '?' operator was completely missing a typecheck for the atype. The
check must make sure that neither the ctype nor the atype match the empty
word.

To allow constructs like (store re)?, we allow the atype to match the empty
word, as long as the lens inside the '?' consumes the value of the current
node, since that also tells as whether to use the inner lens or not.

Various existing lenses failed to typecheck after this change and have been
adapted; this also fixes a bug in the Fstab.lns where existing whitespace
was being replaced with a default tab. The Interfaces.lns needed a fairly
invasive change, and as a consequence produces way more (too many?)
anonymous tree nodes.

Added tests to verify the typechecking of '?'