Complain when nodes in the tree can not be put
When nodes are added to the tree that do not match the appropriate
key/label lens, we used to silently drop them if the key/label was inside a
repetition inside a concat.

Now, we check when we split the tree for a concat that we have used all the
nodes in the tree that we were supposed to, and complain otherwise.

This also uncovered that the test rec-hosts-add.rb was broken, but still
passed.