transform_save: make saving idempotent
When a file needs to be saved, its contents are first generated in a temp
file which is then moved over the original file. Between generating the
temp file and moving it in its final place, read its contents back in and
compare with the contents of the original file. Only move the tempfile over
the original if the contents do indeed differ. This avoids spurious changes
in mtime.