Make finding the parent and siblings of a node uniform
All nodes (including the root) now have a parent node, so
that the start of the list of siblings can be found as
tree->parent->children for any node.

All 'standalone' trees now have a fake root, called 'origin' whose
children are the real root nodes. Another way to look at this is
that the tree is now edge-labeled.