Sudoers: allow backslashes inside sto_to_com and sto_to_com_user
This patch fixes ticket #60 which reports the possible usage of
backslashes in sudoers aliases (for use with e.g. Windows resources).

Backslashes are now allowed only in the middle of sto_to_com and
sto_to_com_cmnd.

Additionally, add a line with backslashes in test_sudoers to prevent
regressions.