tree handling: propagate dirty flag into root
We had an off-by-one error in tree_mark_dirty that lead to the root of the
tree not being marked as dirty. As a consequence, performing a tree_clean
on such a tree would not mark the tree as clean; that caused an aug_load
directly followed by aug_save to write to disk, even though the tree was
unchanged.

  * src/augeas.c (tree_mark_dirty): propagate dirty flag to the root
  * tests/test-load.c (testLoadSave): test that simple load + save
    does not touch the filesystem

This bug was introduced in commit 7278965b