Parse more directives in xorg.conf
The default xorg.conf generated by pyxf86config on Red Hat/Fedora systems
is not parsed by the current xorg.conf. Specifically it does not
understand:

* Screen, in ServerLayout
* InputDevice, in ServerLayout
* SubSection "Display" ..., in Screen

This patch adds support for those values, so that the default xorg.conf on
Red Hat/Fedora systems now parses.

It should also add support for all directives in the Display subsection,
however they are not covered by the test.