Require quotes around strings in xorg.conf
man xorg.conf appears to support this:
           String      a string enclosed in double quote marks (")

This change means that explicitly parsed values will remove quotes in the
get directory and insert them in the put direction.

Currently, only the following entries are explicitly parsed:
* Option
* Screen
* InputDevice
* Driver
* SubSection
* Identifier
* VideoRam
* DefaultDepth
* Device
* All entries in the Display Subsection

The lens will also accept any other entry name not in this list. However,
as it does not know what types to expect, it stores a bare string. This
will include quote marks.