Transparent save of cross-device links
When saving to a file, we first write to a file with extension .augnew and
then rename that file to the actual file. This fails when the original is a
symlink to a file on another device.

We now write the .augnew file next to the canonicalized original file to
avoid that headache. There are still situations where the rename fails,
e.g. when the original file is bindmounted - in that situation, we have to
copy.