Switch to C locale on API entry on systems that have uselocale
Since we need to do all our operations (especially regexp matching) in the
C locale, we switch to that on API entry, and switch back to the user's
locale on API exit.

That also makes it necessary that aug_init wraps the bulk of its work
inside an api_entry/api_exit.

Fixes ticket #35 for systems that have uselocale