Syntax for recursive lenses
Add a 'let rec' construct; the syntax should be considered experimental and
might be changed to some other form of expressing recursion.

With recursive lenses, we can't do the bottom-up typechecking that we do
for regular lenses; we need to consider the whole lens at once. This is
achieved by inserting a callback to lns_check_rec when we parse 'let rec
ident = exp'.

  * src/builtin.c (lns_check_rec_glue): new function
  * src/lexer.l: new token LETREC
  * src/parser.y: add syntax rule for LETREC and desugar it into a call to
    lns_check_rec
  * src/syntax.h: make free_term available, constant for name
    of let_check_rec within the interpreter
  * src/syntax.c (free_term): now extern, not static