* src/jmt.c (jmt_visit): allow for toplevel lenses that are not L_REC
To reconstruct the parse of a lens that is not L_REC, but uses L_REC,
e.g. l* where l is L_REC, we can not assume any longer that the root of the
parse tree has only one child. Consequently, we need to start
reconstruction with listing all the children of the root node.

Split build_tree into build_children (collects a list of child nodes) and
build_tree which creates the tree for a nonterminal.

To start reconstruction, we now start with collecting the list of children
of the root of the parse tree; the Earley graph does not contain an item
for that, which makes this operation slightly special.