Properly handle lenses that use L_REC lenses
In constructs involving a recursive lens, we need to take the body of the
recursive lens into consideration. We avoid infinite recursion by stopping
when we see an L_REC that is marked as rec_internal, since that's the
struct lens that points back to the top of the lens from within it.

jmt_build now expects to be handed a properly set up recursive lens, which
means that for an L_REC both the internal and external instances have to
exist and be connected

  * src/lens.c (lns_check_rec): defer building the jmt until the end
  * src/jmt.c: when building the jmt, consider L_REC as long as it's not
    internal