jmt: distinguish between successful parse and syntax error at end of input
If a syntax error happened after the last character of the input string
(e.g., becuase the input was incomplete), rec_process would think the parse
was successful, and then notice that there was no parse result. That would
lead to an incorrect error message hinting at an internal error, when it
was really a syntax error.

By changing the return value for jmt_visit, it's now possible to
distinguish between internal errors, syntax errors and successful parses
properly.