Debctrl (multi_line_entry): fix bad regexp
The lens 'del /^ / " "' is not valid, since we do not allow '^' as a meta
character. Changing that to 'del / / " "' exposed a typecheck problem in
lns: both multi_line_entry and eol match blank lines. Fixed this by
requiring that a multi_line_entry has to have at least one non-blank
character.