Sysconfig: new lens for a subdialect of shell
Using shell files with simple variable assignments is a very popular config
file format. The Shellvars lens doesn't serve those files perfectly well,
since it tries to preserve details important for general shell scripts that
are ignored by sysconfig files.

This lens is a variation on Shellvars that changes a few important aspects
of it:

  * Strip quotes from values when reading a file, and restore quotes as
    needed upon writing
  * Pretend there's no difference between single and double quotes (this
    assumption makes it possible to add quotes as needed solely based on
    syntax)
  * Do not support shell arrays or backticks; support for these could be
    restored by somebody sufficiently motivated, though neither of them
    should be used in a sysconfig file

The lens doesn't autoload any files yet, though we should consider moving
most of the files that Shellvars loads over to this lens.