Properly record the new mtime of a saved file in the tree
transform_save used to pass an incorrect file name to add_file_info (the
name was missing the root prefix), leading to an mitme of 0 for all saved
files. That in turn causes freshly saved files to be unnecessarily reloaded
on the next aug_load.

  * src/transform.c (mtime_as_string): tolerate a NULL filename, and return
    an mtime of 0; (transform_save): pass the full file name, including the
    root to add_file_info
  * tests/test-load.c (testReloadDeletedMeta): the test is now
    time-sensitive; to work around that, set the mtime of the changed tree
    to the original mtime of the file
  * tests/test-save.c (testMtime): verify that we record the actual mtime
    and not just 0 in the tree after a save.

Thanks to Laine Stump for spotting this