Preserve parse errors under /augeas//error
With the changes in commit 5ee81630, a subtree in /augeas/files that is
dirty after trying to load all files indicates a file that was loaded
before, but is not loaded anymore (e.g., because the set of lenses changed
since the last time we loaded) In that situation, both the subtree in
/augeas/files as well as the corresponding subtree in /files is deleted.

The code in store_error, which is called by load_file when a parse error
happens, would leave that subtree in /augeas/files dirty. This commit fixes
that by making sure we mark the error entries as clean. It also revamps
store_error to use the lighter-weight tree_xxx functions instead of the
aug_xxx functions.

Fixes bug #138