Add square lens to language
The new primitive square lens allows to handle XML like open and close tags.
It takes 2 arguments, a regexp for the tag and a lens for the content. The
square lens verify that open and close tags are matched in the get direction,
and put direction handles the close tag by copying the key.

  * src/builtin.c: add "square" keyword to language
  * src/lens.c: lns_make_square to compile the lens from actual simpler
    lenses
  * src/lens.h: append L_SQUARE to lens_tag enum to identify the square lens
    type object
  * src/jmt.c: handle square lens in parser
  * src/get.c: add char* square to state struct to hold the last L_DEL in
    square to perform syntax check, and add all get handling functions for
    L_SQUARE
  * src/put.c: modify create_del to render current key when default string is
    NULL and add all put handling functions
  * tests/modules/pass_square.aug: unit tests to validate the square lens,
    regular and recursives