Add xml lens
The XML lens parse arbitrary XML documents. Most, but not all features of XML
standard are implemented, partially because the nature of augeas. Tests comes
from OASIS test suite.

Supported:
  * Document type declaration: elements, attlist, entity
  * XML Prolog
  * XML comments
  * XML processing instructions
  * Neested elements, including empty elements
  * Element attributes
  * Mixed content, text and elements
  * CDATA sections

Not supported:
  * Single quotes attributes: double quotes is mandatory
  * External entity reference, because augeas can't load a related file within a lens
  * Document validation: augeas don't enforce DTD

This lens doesn't respect the PutGet law, becuse of an ambiguity in union.put
with text nodes. If a text node is created as a sibling of another text node, a
put and a get will merge those two nodes. In practice, no error occur and this
is a desirable behavior.