Replace augtest by test-augtool.sh
This removes the silly dependency on Ruby, since augtest was the only
Ruby script in the whole source tree.

The test files for test-augtool.sh live in tests/test-augtool, and are
functionally identical to the Ruby test files tests/*.rb. The only
exception is that test-root.rb has been removed - it tested something that
is better tested in test-xpath.c

The new test-augtool.sh test is also a little more efficient in that it
only copies the file needed by a test into the root, and only sets the lens
for that file up for loading. This reduces the runtime of the tests from ~
30s on my machine to under 2s.