Add aug_span API function V2
aug_span API function provides information about the node of the specified
path. It sets the absolute path of the file name and spans of label, value
and span inside the node. The span includes all matching chars of the actual
lens and it's descendant. Hence, the top level node will span the entire file.
It returns 0 in case of success and -1 if the node is not associated to a file
or if the path is invalid.  The main use case is to make possible to display
the related file with the node elements hightlighted in a UI.

The API option AUG_NO_NODE_SPAN provided to aug_init disable the allocation
structures to keep node spans. It saves about 5% of memory for a typical
tree.

  * src/augeas.c : add aug_span implementation and option handling to aug_in
    according to the option /augeas/span in the tree
  * src/augeas.h : add declaration of aug_span and constants for option
  * src/augeas_sym.version : make the new function visible in libaugeas
  * src/augtool.c : add the "nodepos" command to augtool
  * src/get.c : gather filename, label, value and span spans if any
  * src/info.c : create and release functions for struct node_span
  * src/info.h : add definition of struct node_info
  * src/internal.h : add node_info to struct tree
  * src/transform.c : duplicate filename string
  * tests/test-load.c : 3 basic unit tests for general correct behavior