Fix unmatch bracket causes segfault in augtool.c
When command is parsed in augtool, function nexttoken verify that path has
valid and matching brackets and return NULL if not. The function cleanpath that
follows was not handling the NULL case, leading to a segmentation fault. By
checking for NULL value, the error is thrown adequately.

Fix bug #186

Thanks to spamaps for the help to resolve this issue.