path expressions: parse [expr1 or expr2] correctly
Since commit ea010d85, we would interpret the predicate [e1 or e2] as
matching any node with a child labelled 'e1 or e2'; this patch restores the
semantics that the predicate is true if either e1 or e2 are true. It does
the same for [e1 and e2]

Of course, this is a heinous hack.