Fix regular expression escaping
This patch fixes:

  /[\/]/

Previously, this matched either a backslash or a slash, since \/ was not
an escape sequence known to unescape(). Now this matches only a slash.

We pass an extra argument to escape() and unescape() listing the extra
characters that should be handled. STR_ESCAPES and RX_ESCAPES provide
the extra characters for strings and regexps respectively, and NULL may
be passed if no extra characters should be handled.

Signed-off-by: Michael Chapman <mike@very.puzzling.org>