compile_exp: don't return an uninitialized pointer upon failure
As a result of its improved flow analysis gcc-4.7.0 20111202 reported:

  syntax.c: In function 'compile_exp':
  syntax.c:1694:5: warning: 'v' may be used uninitialized in this \
  function [-Wmaybe-uninitialized]

However, compile_exp calls many functions that return values that are
then assigned to "v".  Only two of those had a problem: in each there
is an error path by which the returned pointer is not initialized.
* src/syntax.c (compile_concat): Initialize "v".
(compile_minus): Likewise.