Case-insensitive regexps: fix a problem with number of groups
When we expand a regexp to a case-sensitive regexp in regexp_expand_nocase,
it is possible that the number of groups in that regexp changes,
e.g. /(a)|(b)/i becomes /[AaBb]/. That confuses the parser when that regexp
is used in a union or concatenation, since it expects the match following
the match for this regexp to be in group 3, not group 1.