Fix bug #293 primary statement in dhcpd.conf
The "primary" statement does not support any arguments, while it should when
it's used in the "zone" block statement. The solution is to add the primary
statement in the statements that requires a string argument. The primary
statement can thus take a value or not, and thus do not cause any ambiguity in
the put direction.