Revert f04d7dfb3fe7c54b0be650e68aa7778eaf02ea66
This was a subtle feature which uses counters named after the setting key to
index the values beneath it, over all instances of the setting key.  This allows
for path expressions such as /files/etc/ssh/sshd/AllowEnv/10 to map to the
tenth value, irrespective of how many setting lines they are split over.