Correct locations table for transform_save()
Commit 16387744 introduced a table with locations of temporary, destination and
backup file in various modes as part of the comment for transform_save().  This
corrects two incorrectly listed temporary file locations.

Note that PATH_canon is somewhat misleading, as canonicalized path is always
used when not using NEWFILE save flag, while it is meant to indicate whether
file is created in a directory that contains symlink, or a directory that
contains symlink's target.