dhcpd, adding support for set command
This allows you to define variables in your dhcp configuration.  They
can be arbitrarily complicated.

Things like this:

      set ClientIP = binary-to-ascii(10, 8, ".", leased-address);
      set ClientMac = binary-to-ascii(16, 8, ":" , substring(hardware, 1, 6));

I also added most of the rest of the functions from dhcp-eval and listed
the limitations of the current parser.  I don't think I'm going to be
able to fix most of it but I think I'm getting into obscure features
that most people don't bother using.

At this point if they need to do very strange class maps they could put
them in an include file and not have augeas parse it.