Fix for readline add_history with interactive mode
If you use interactive mode, augtool never adds anything to history that
you type once the input becomes a normal tty.

This is because it was checking stdin, which as far as the system is
concerned is still !isatty().  I'm not sure of the better way to fix
this.  An freopen stdin with /dev/tty might be good, but just changing
the add_history check to use rl_instream seems to work fine as long as
it doesn't cause unintended effects.