Analyze 'short iteration' errors better
When an iteration fell short, we would just throw our hands in the air and
produce not very helpful information. We now try to find the exact place
where some part of the iterated lens still matched and the rest didn't and
provide that information to the user.

Note that these errors will only be produced for regular lenses, not for
recursive lenses