Suppress repeated errors about not finding a module
We used to print 'Could not load module XYZ' every time that module wsa
mentioned. To cut down the number of annoying error messages, we now create
an empty dummy module when we fail to look one up; subsequent lookups will
find that and therefore not print the error message again.